[SPARK-15686][SQL] Move user-facing structured streaming classes into sql.streaming
